backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==23.0.0
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
asyncpg==0.29.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
